requirements.txt
# Core dependencies
pillow>=10.0.0
numpy>=1.24.0

# OCR Models
pytesseract>=0.3.10
paddlepaddle>=2.5.0
paddleocr>=2.7.0
surya-ocr>=0.4.0

# LLM clients (at least one required for LLM reconciliation)
openai>=1.0.0

# Optional: Other LLM clients
# anthropic>=0.7.0
# google-generativeai>=0.3.0

# Development (optional)
# pytest>=7.0.0
# black>=23.0.0
